requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==43.0.0
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.7
pillow==10.4.0
pycparser==2.22
PyJWT==2.9.0
PyMySQL==1.1.1
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.1
stripe==10.6.0
typing_extensions==4.12.2
urllib3==2.2.2
